+++ NEWS.md
@@ -7,7 +7,7 @@
   See [Posit Cloud's Announcement](https://posit.cloud/learn/whats-new#publishing)
   for more information and to request access.
 
-* Update company and product name for rebranding to Posit.
+* Update company and product names for rebranding to Posit.
 
 ## 0.8.28